requirements.txt
Pympler==0.4.3
python-magic==0.4.12
